requirements.txt
asgiref==3.9.1
certifi==2025.8.3
cffi==1.17.1
charset-normalizer==3.4.2
cryptography==45.0.5
Django==5.2.4
django-allauth==65.10.0
django-cors-headers==4.7.0
django-filter==25.1
djangorestframework==3.16.0
idna==3.10
oauthlib==3.3.1
pycparser==2.22
PyJWT==2.10.1
python-dotenv==1.1.1
requests==2.32.4
requests-oauthlib==2.0.0
sqlparse==0.5.3
tzdata==2025.2
urllib3==2.5.0
<<<<<<< HEAD

# PostgreSQL
psycopg2-binary==2.9.10

# Data Processing
pandas==2.1.4
numpy==1.26.4
openpyxl==3.1.5

# Web Scraping (추후 설치 예정)
# beautifulsoup4==4.12.2
# selenium==4.15.2
# lxml==4.9.3

# Environment Variables
# python-decouple==3.8
=======
gunicorn==21.2.0
psycopg2-binary
>>>>>>> f951233ffdadc0dfc6c3c39d0499cd9473321ba2
